+++ packages/dara-components/tests/python/common/test_select.py
@@ -20,18 +20,12 @@ class TestSelectComponent(unittest.TestCase):
         expected_dict = {
             'name': 'Select',
             'props': {
-                'bold': False,
                 'children': [],
-                'italic': False,
                 'items': [
                     {'label': 'test1', 'value': 'test1'},
                     {'label': 'test2', 'value': 'test2'},
                     {'label': 'test3', 'value': 'test3'},
                 ],
-                'max_rows': 3,
-                'multiselect': False,
-                'searchable': False,
-                'underline': False,
                 'value': value.dict(exclude_none=True),
             },
             'uid': str(test_uid),