+++ packages/dara-components/tests/python/common/test_table.py
@@ -38,10 +38,6 @@ class TestTableComponent(unittest.TestCase):
             'name': 'Table',
             'props': {
                 'children': [],
-                'bold': False,
-                'include_index': True,
-                'italic': False,
-                'underline': False,
                 'columns': [
                     {
                         'col_id': 'name',
@@ -56,10 +52,6 @@ class TestTableComponent(unittest.TestCase):
                     },
                 ],
                 'data': data_var.dict(exclude_none=True),
-                'multi_select': False,
-                'show_checkboxes': True,
-                'searchable': False,
-                'suppress_click_events_for_selection': False,
             },
             'uid': 'uid',
         }
@@ -78,10 +70,6 @@ class TestTableComponent(unittest.TestCase):
             'name': 'Table',
             'props': {
                 'children': [],
-                'bold': False,
-                'italic': False,
-                'include_index': True,
-                'underline': False,
                 'columns': [
                     {
                         'col_id': 'col1',
@@ -96,10 +84,6 @@ class TestTableComponent(unittest.TestCase):
                     {'col_id': 'col2', 'label': 'col2'},
                 ],
                 'data': data_var.dict(exclude_none=True),
-                'multi_select': False,
-                'show_checkboxes': True,
-                'searchable': False,
-                'suppress_click_events_for_selection': False,
             },
             'uid': 'uid',
         }