+++ packages/dara-components/tests/python/common/test_text.py
@@ -14,13 +14,8 @@ class TestTextComponent(unittest.TestCase):
         expected_dict = {
             'name': 'Text',
             'props': {
-                'align': 'left',
-                'bold': False,
                 'children': [],
-                'formatted': False,
-                'italic': False,
                 'text': 'Some Text',
-                'underline': False,
             },
             'uid': 'uid',
         }
@@ -32,10 +27,6 @@ class TestTextComponent(unittest.TestCase):
             'props': {
                 'children': [],
                 'align': 'center',
-                'bold': False,
-                'formatted': False,
-                'italic': False,
-                'underline': False,
                 'width': '10px',
                 'text': 'Some Text',
             },