+++ packages/dara-components/tests/python/plotting/test_bokeh.py
@@ -27,9 +27,6 @@ class TestBokehComponent(unittest.TestCase):
             'name': 'Bokeh',
             'props': {
                 'document': cmp.document,
-                'bold': False,
-                'italic': False,
-                'underline': False,
             },
             'uid': str(test_uid),
         }