+++ packages/dara-core/tests/python/test_components.py
@@ -64,18 +64,12 @@ def test_children():
         'uid': instance.uid,
         'props': {
             'foo': 'bar',
-            'underline': False,
-            'bold': False,
-            'italic': False,
             'children': [
                 {
                     'name': 'TestComponent',
                     'uid': instance.children[0].uid,
                     'props': {
                         'foo': 'baz',
-                        'bold': False,
-                        'underline': False,
-                        'italic': False,
                     },
                 },
             ],
@@ -128,7 +122,6 @@ def test_fallback():
                 'fallback': jsonable_encoder(fallback),
                 'func_name': 'TestComponent',
                 'dynamic_kwargs': {},
-                'polling_interval': None,
             },
             'uid': component.uid,
         }